[ci] Create custom image for local Jenkins agent
Here is a list of requirements that the local Jenkins agent should have:

- It should have SSH enabled in order to be connected from the
controller.
- It should have Docker installed to run ephemeral containers for each
Jenkins build.

Therefore, a custom Containerfile is written based on docker:dind image
(Docker in Docker).
On top of this image, the tools below are installed:

- Java (requirement by the controller, since an agent executable is
basically a .jar file)
- OpenSSH (the controller will connect to the agent via SSH when needed)

And finally, here is a list of configuration that is done in this image:
- A new user called 'jenkins' is created with necessary directories and
permissions.
- Directories and necessary keys are generated for OpenSSH. Also,
password authentication is turned off.
- The entrypoint of the image is changed to include sshd along with
dockerd.